First attempt, doesn't work yet
This experimental code already contains most of the initial ideas,
and quite a number of fixes found neccessary along the way.

For example, fakechroot does treat symblinks which already exist at startup
as links pointing outside the directory tree, so we need to relativize
all symlinks.

Nevertheless, in practice things fail with

    Unable to connect to host 127.0.0.1 on port 7055 after 45000 ms.

For details and discussion see the issue reports.